Refactor resource.rb
